Revert "Webserver changes by changes.txt"
