--- ton-test-liteclient-full/lite-client-with-webserver/.gitignore
@@ -1,5 +1,0 @@
-.idea/
-cmake-build-debug/
-cmake-build-release/
-**/.DS_Store
-contrib/
