--- ton-test-liteclient-full/lite-client-with-webserver/incl/webserver/test_node.hpp
@@ -1,232 +1,0 @@
-#pragma once
-
-#include <ton/ton-tl.hpp>
-#include <crypto/block/block.h>
-#include <td/utils/filesystem.h>
-#include <ton/ton-shard.h>
-
-#include "ton/ton-types.h"
-#include "terminal/terminal.h"
-#include "vm/cells.h"
-#include "adnl/adnl-ext-client.h"
-#include "tl-utils/tl-utils.hpp"
-#include "auto/tl/ton_api_json.h"
-
-#include "td/utils/OptionsParser.h"
-#include "td/utils/Time.h"
-#include "td/utils/filesystem.h"
-#include "td/utils/format.h"
-#include "td/utils/Random.h"
-#include "td/utils/crypto.h"
-#include "td/utils/port/signals.h"
-#include "td/utils/port/stacktrace.h"
-#include "td/utils/port/StdStreams.h"
-#include "td/utils/port/FileFd.h"
-
-#include "terminal/terminal.h"
-#include "ton/ton-tl.hpp"
-#include "block/block-db.h"
-#include "block/block.h"
-#include "block/block-auto.h"
-#include "block/mc-config.h"
-#include "vm/boc.h"
-#include "vm/cellops.h"
-#include "vm/cells/MerkleProof.h"
-#include "ton/ton-shard.h"
-
-#include <boost/property_tree/ptree.hpp>
-#include <boost/property_tree/json_parser.hpp>
-
-#define NEW_WEBSERVER
-
-
-#ifdef NEW_WEBSERVER
-    #include "new_http_server.h"
-    using namespace SimpleWeb;
-    using HttpServer=CustomHttpServer;
-#else
-    #include "http_server.hpp"
-    using HttpServer = SimpleWeb::Server<SimpleWeb::HTTP>;
-#endif
-
-// Short alias for this namespace
-namespace pt = boost::property_tree;
-
-
-using td::Ref;
-class Responce;
-
-class TestNode : public td::actor::Actor {
-public:
-    TestNode() {
-    }
-    ~TestNode() = default;
-
-    TestNode(const TestNode&)= delete;
-    TestNode& operator=(const TestNode&)= delete;
-    TestNode(const TestNode&&)= delete;
-
-
-    void run();
-
-    void conn_ready() {
-        LOG(ERROR) << "conn ready";
-        ready_ = true;
-        if (!inited_) {
-            run_init_queries();
-        }
-    }
-    void conn_closed() {
-        ready_ = false;
-    }
-    void set_local_config(std::string str) {
-        local_config_ = str;
-    }
-    void set_global_config(std::string str) {
-        global_config_ = str;
-    }
-    void set_db_root(std::string db_root) {
-        db_root_ = db_root;
-    }
-    void set_readline_enabled(bool value) {
-        readline_enabled_ = value;
-    }
-    void set_liteserver_idx(td::int32 idx) {
-        liteserver_idx_ = idx;
-    }
-    void set_update_on_demand(bool value) {
-        update_on_demand_enabled_ = value;
-    }
-
-    void start_up() override {
-    }
-    void tear_down() override {
-        // FIXME: do not work in windows
-        //td::actor::SchedulerContext::get()->stop();
-    }
-
-    bool envelope_send_query(td::BufferSlice query, td::Promise<td::BufferSlice> promise);
-    bool envelope_send_web(td::BufferSlice query, td::Promise<td::BufferSlice> promise,
-            std::shared_ptr<Response> response);
-    void parse_line(td::BufferSlice data);
-
-    // web server methods
-    void get_server_time_web(std::shared_ptr<Response> response);
-    void get_account_state_web(std::string address, std::shared_ptr<Response> response);
-    void got_account_state_web(ton::BlockIdExt blk, ton::BlockIdExt shard_blk, td::BufferSlice shard_proof,
-                               td::BufferSlice proof, td::BufferSlice state, ton::WorkchainId workchain,
-                               ton::StdSmcAddress addr, std::shared_ptr<Response> response);
-    void get_block_web(std::string blkid_str, std::shared_ptr<Response> response, bool dump = true);
-    void got_block_web(ton::BlockIdExt blkid, td::BufferSlice data, bool dump, std::shared_ptr<Response> response);
-    bool give_block_header_description(std::ostringstream& out, ton::BlockIdExt blkid, Ref<vm::Cell> root, int mode);
-
-    bool get_server_mc_block_id_web(std::shared_ptr<Response> response);
-
-
-
-    // Web Server Methods
-    static void run_web_server(td::actor::Scheduler* scheduler, td::actor::ActorOwn<TestNode>* x);
-    static void web_error_response(std::shared_ptr<Response> response, std::string msg);
-    static void web_success_response(std::shared_ptr<Response> response, std::string msg);
-    static void web_success_response(std::shared_ptr<Response> response, pt::ptree root);
-
-    void web_last(){
-        get_server_mc_block_id_silent();
-    }
-private:
-    void run_init_queries();
-    bool get_server_time();
-    bool get_server_mc_block_id();
-    bool get_server_mc_block_id_silent();
-    void got_server_mc_block_id(ton::BlockIdExt blkid, ton::ZeroStateIdExt zstateid);
-    void got_server_mc_block_id_silent(ton::BlockIdExt blkid, ton::ZeroStateIdExt zstateid);
-    bool request_block(ton::BlockIdExt blkid);
-    bool request_state(ton::BlockIdExt blkid);
-    void got_mc_block(ton::BlockIdExt blkid, td::BufferSlice data);
-    void got_mc_state(ton::BlockIdExt blkid, ton::RootHash root_hash, ton::FileHash file_hash, td::BufferSlice data);
-    td::Status send_set_verbosity(std::string verbosity);
-    td::Status send_ext_msg_from_filename(std::string filename);
-    td::Status save_db_file(ton::FileHash file_hash, td::BufferSlice data);
-    bool get_account_state(ton::WorkchainId workchain, ton::StdSmcAddress addr, ton::BlockIdExt ref_blkid);
-    void got_account_state(ton::BlockIdExt ref_blk, ton::BlockIdExt blk, ton::BlockIdExt shard_blk,
-                           td::BufferSlice shard_proof, td::BufferSlice proof, td::BufferSlice state,
-                           ton::WorkchainId workchain, ton::StdSmcAddress addr);
-    bool get_all_shards(bool use_last = true, ton::BlockIdExt blkid = {});
-    void got_all_shards(ton::BlockIdExt blk, td::BufferSlice proof, td::BufferSlice data);
-    bool get_block(ton::BlockIdExt blk, bool dump = false);
-    void got_block(ton::BlockIdExt blkid, td::BufferSlice data, bool dump);
-    bool get_state(ton::BlockIdExt blk, bool dump = false);
-    void got_state(ton::BlockIdExt blkid, ton::RootHash root_hash, ton::FileHash file_hash, td::BufferSlice data,
-                   bool dump);
-    bool get_block_header(ton::BlockIdExt blk, int mode);
-    void got_block_header(ton::BlockIdExt blkid, td::BufferSlice data, int mode);
-    bool show_block_header(ton::BlockIdExt blkid, Ref<vm::Cell> root, int mode);
-    bool show_state_header(ton::BlockIdExt blkid, Ref<vm::Cell> root, int mode);
-    bool get_one_transaction(ton::BlockIdExt blkid, ton::WorkchainId workchain, ton::StdSmcAddress addr,
-                             ton::LogicalTime lt, bool dump = false);
-    void got_one_transaction(ton::BlockIdExt req_blkid, ton::BlockIdExt blkid, td::BufferSlice proof,
-                             td::BufferSlice transaction, ton::WorkchainId workchain, ton::StdSmcAddress addr,
-                             ton::LogicalTime trans_lt, bool dump);
-    bool get_last_transactions(ton::WorkchainId workchain, ton::StdSmcAddress addr, ton::LogicalTime lt,
-                               ton::Bits256 hash, unsigned count, bool dump);
-    void got_last_transactions(std::vector<ton::BlockIdExt> blkids, td::BufferSlice transactions_boc,
-                               ton::WorkchainId workchain, ton::StdSmcAddress addr, ton::LogicalTime lt,
-                               ton::Bits256 hash, unsigned count, bool dump);
-
-    bool do_parse_line();
-    bool show_help(std::string command);
-    std::string get_word(char delim = ' ');
-    int skipspc();
-    std::string get_line_tail(bool remove_spaces = true) const;
-    bool eoln() const;
-    bool seekeoln();
-    bool set_error(td::Status error);
-    bool set_error(std::string err_msg);
-    void show_context() const;
-    bool parse_account_addr(ton::WorkchainId& wc, ton::StdSmcAddress& addr);
-    static int parse_hex_digit(int c);
-    static bool parse_hash(const char* str, ton::Bits256& hash);
-    static bool parse_uint64(std::string word, td::uint64& val);
-    bool parse_hash(ton::Bits256& hash);
-    bool parse_lt(ton::LogicalTime& lt);
-    bool parse_block_id_ext(ton::BlockIdExt& blkid, bool allow_incomplete = false);
-    bool parse_block_id_ext(std::string blk_id_string, ton::BlockIdExt& blkid, bool allow_incomplete = false) const;
-    bool register_blkid(const ton::BlockIdExt& blkid);
-    bool show_new_blkids(bool all = false);
-    bool complete_blkid(ton::BlockId partial_blkid, ton::BlockIdExt& complete_blkid) const;
-
-    static bool parse_account_addr(std::string acc_string, ton::WorkchainId& wc, ton::StdSmcAddress& addr);
-
-private:
-    std::string local_config_ = "ton-local.config";
-    std::string global_config_ = "ton-global.config";
-
-    td::actor::ActorOwn<ton::AdnlExtClient> client_;
-    td::actor::ActorOwn<td::TerminalIO> io_;
-
-    bool readline_enabled_ = true;
-    td::int32 liteserver_idx_ = -1;
-
-    bool ready_ = false;
-    bool inited_ = false;
-    bool update_on_demand_enabled_ = true;
-    std::string db_root_;
-
-    int server_time_ = 0;
-    int server_time_got_at_ = 0;
-
-    ton::ZeroStateIdExt zstate_id_;
-    ton::BlockIdExt mc_last_id_;
-
-    ton::BlockIdExt last_block_id_, last_state_id_;
-    td::BufferSlice last_block_data_, last_state_data_;
-
-    std::string line_;
-    const char *parse_ptr_, *parse_end_;
-    td::Status error_;
-
-    std::vector<ton::BlockIdExt> known_blk_ids_;
-    std::size_t shown_blk_ids_ = 0;
-
-    std::unique_ptr<ton::AdnlExtClient::Callback> make_callback();
-};
