+++ ton-test-liteclient-full/lite-client-with-webserver/src/web_server/status_code.hpp
@@ -1,4 +1,5 @@
-#pragma once
+#ifndef SIMPLE_WEB_STATUS_CODE_HPP
+#define SIMPLE_WEB_STATUS_CODE_HPP
 
 #include <cstdlib>
 #include <map>
@@ -171,3 +172,5 @@ namespace SimpleWeb {
     return pos->second;
   }
 } // namespace SimpleWeb
+
+#endif // SIMPLE_WEB_STATUS_CODE_HPP
