--- ton-test-liteclient-full/lite-client-with-webserver/tests/mock_api.cpp
@@ -1,18 +1,0 @@
-#include "mock_api.hpp"
-
-MockApi::MockApi(){
-	
-}
-
-
-void mockGetFoo(){
-	
-}
-
-void web_error_response(std::shared_ptr<SimpleWeb::Response> response, std::string msg){
-	
-}
-
-void web_success_response(std::shared_ptr<SimpleWeb::Response> response, std::string msg){
-	
-}
